requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
jinja2==3.1.3
python-multipart==0.0.9
openpyxl==3.1.2
python-dotenv==1.0.1
aiofiles==23.2.1
